Move OpenMandriva to DistributionInstaller
